requirements.txt
psycopg2-binary
google-cloud-aiplatform
requests 